New Character Indicator
There was not an indication to the user when a new character was
generated. This meant that if the same character was generated twice in
a row, the user wouldn't know if it was the same observation
in the sequence or a new one.

Added a consistent string (---) to indicate that a new character has
been generated.